# Speed up interpolation by implementing it in C++