--- a/setup.py
+++ b/setup.py
@@ -56,6 +56,7 @@ pybind_extensions = [
             # Path to pybind11 headers
             get_pybind_include(),
         ],
+        extra_compile_args=['-std=c++11'],
         language='c++'
     ),
 ]