--- a/setup.py
+++ b/setup.py
@@ -56,7 +56,7 @@ pybind_extensions = [
             # Path to pybind11 headers
             get_pybind_include(),
         ],
-        extra_compile_args=['-std=c++11'],
+        extra_compile_args=['-std=c++17'],
         language='c++'
     ),
 ]