--- a/tools/RAiDER/__init__.py
+++ b/tools/RAiDER/__init__.py
@@ -1,0 +1,13 @@
+"""
+Raytracing Atmospheric Delay Estimation for RADAR
+
+Copyright (c) 2019-2020, California Institute of Technology ("Caltech"). All rights reserved.
+"""
+
+# Import compiled modules so we can do something like
+#
+# import RAiDER
+# RAiDER.Geo2rdr
+from RAiDER import Geo2rdr, interpolate, makePoints
+
+__copyright__ = 'Copyright (c) 2019-2020, California Institute of Technology ("Caltech"). All rights reserved.'