@dbekaert 2020-07-17
@Askaholic this looks very promising! This optimization is definitely paying off.